@@ -1,10 +1,15 @@
-import {  flags, SfdxCommand } from '@salesforce/command';
-import { Messages} from '@salesforce/core';
+import { Flags } from '@oclif/core';
+import { SfCommand, Flags as SfFlags } from '@salesforce/sf-plugins-core';
+import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
-import { createTmpDir } from '../../../../utils/createTmpDir';
-const fse = require('fs-extra');
-const compressing = require('compressing');
-const fs = require('fs');
+import { createTmpDir } from '../../../../utils/createTmpDir.js';
+import gracefulFs from 'graceful-fs';
+import compressing from 'compressing';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -13,47 +18,40 @@ Messages.importMessagesDirectory(__dirname);
 // or any library that is using the messages framework can also be loaded this way.
 const messages = Messages.loadMessages('vnjikedx', 'org');
 
-export default class Upsert extends SfdxCommand {
-  public static description = messages.getMessage('commandDescription');
+export default class VnjikeMetadataLabelUpsert extends SfCommand<AnyJson> {
+  public static readonly summary = messages.getMessage('commandDescription');
+  public static readonly description = messages.getMessage('commandDescription');
 
-  public static examples = [
-    `$ vnjike:metadata:label:upsert --targetusername myOrg@example.com --targetlabelname "apiname" --targetlabelvalue "value"
-    $ vnjike:metadata:label:upsert -u myOrg@example.com -n "apiname" -v "value"
-  `
+  public static readonly examples = [
+    `<%= config.bin %> <%= command.id %> --target-org myOrg@example.com --target-label-name "apiname" --target-label-value "value"`,
+    `<%= config.bin %> <%= command.id %> -o myOrg@example.com -n "apiname" -v "value"`
   ];
 
-  public static args = [{ name: 'file' }];
-
-  protected static flagsConfig = {
-    // flag with a value (-n, --name=VALUE)
-    targetlabelname: flags.string({
+  public static readonly flags = {
+    'target-org': SfFlags.requiredOrg(),
+    'target-label-name': Flags.string({
       required: true,
       char: 'n',
-      description: messages.getMessage('targetlabelname')
+      description: messages.getMessage('targetlabelname'),
+      summary: messages.getMessage('targetlabelname')
     }),
-    targetlabelvalue: flags.string({
+    'target-label-value': Flags.string({
       required: true,
       char: 'v',
-      description: messages.getMessage('targetlabelvalue')
+      description: messages.getMessage('targetlabelvalue'),
+      summary: messages.getMessage('targetlabelvalue')
     })
   };
 
-  // Comment this out if your command does not require an org username
-  protected static requiresUsername = true;
-
-  // Comment this out if your command does not support a hub org username
-  protected static supportsDevhubUsername = false;
-
-  // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
-  protected static requiresProject = false;
-
   public async run(): Promise<AnyJson> {
+    const { flags } = await this.parse(VnjikeMetadataLabelUpsert);
+    
     // evaluate entries
-    const targetlabelname = this.flags.targetlabelname;
-    const targetlabelvalue = this.flags.targetlabelvalue;
+    const targetLabelName = flags['target-label-name'];
+    const targetLabelValue = flags['target-label-value'];
 
-    // this.org is guaranteed because requiresUsername=true, as opposed to supportsUsername
-    const conn = this.org.getConnection();
+    // Get the org connection
+    const conn = flags['target-org'].getConnection();
     const tmpDir = createTmpDir();
 
     try {
@@ -62,52 +60,51 @@ export default class Upsert extends SfdxCommand {
       // create a file
       let packageContent =
         '<?xml version="1.0" encoding="UTF-8"?>\n<Package xmlns="http://soap.sforce.com/2006/04/metadata">\n<types>\n<members>';
-      packageContent += targetlabelname;
+      packageContent += targetLabelName;
       packageContent +=
         '</members>\n<name>CustomLabel</name>\n</types>\n<version>46.0</version>\n</Package>';
-      fs.writeFileSync(tmpDir + '\\package.xml', packageContent);
-      fs.mkdirSync(tmpDir + '\\labels');
+      gracefulFs.writeFileSync(tmpDir + '\\package.xml', packageContent);
+      gracefulFs.mkdirSync(tmpDir + '\\labels');
       const customLabelContent =
         '<?xml version="1.0" encoding="UTF-8"?>\n<CustomLabels xmlns="http://soap.sforce.com/2006/04/metadata">\n<labels>\n<fullName>' +
-        targetlabelname +
+        targetLabelName +
         '</fullName>\n<language>en_US</language>\n<protected>false</protected>\n<shortDescription>' +
-        targetlabelname +
+        targetLabelName +
         '</shortDescription>\n<value>' +
-        targetlabelvalue +
+        targetLabelValue +
         '</value>\n</labels>\n</CustomLabels>';
 
-      fs.writeFileSync(
+      gracefulFs.writeFileSync(
         tmpDir + '\\labels\\CustomLabels.labels',
         customLabelContent
       );
       const zipFile = tmpDir + '.zip';
       await compressing.zip.compressDir(tmpDir, zipFile);
 
-      const zipStream = fs.createReadStream(zipFile);
+      const zipStream = gracefulFs.createReadStream(zipFile);
       const result = await conn.metadata.deploy(zipStream, {});
 
       let done = false;
-
       let deployResult;
 
       while (!done) {
         deployResult = await conn.metadata.checkDeployStatus(result.id);
         done = deployResult.done;
         if (!done) {
-          this.ux.log('Deployement status: ' + deployResult.status);
+          console.log('Deployment status: ' + deployResult.status);
           await new Promise(sleep => setTimeout(sleep, 5000));
         }
       }
 
-      // delete tempory directory & file
-      fse.removeSync(tmpDir);
-      fs.unlinkSync(zipFile);
+      // delete temporary directory & file
+      gracefulFs.rmSync(tmpDir, { recursive: true, force: true });
+      gracefulFs.unlinkSync(zipFile);
       console.log('label deployment finished!');
-    } catch (err) {
-      this.ux.errorJson(err);
+    } catch (err: any) {
+      throw new Error(`Label upsert failed: ${err?.message || 'Unknown error'}`);
     }
 
     // Return an object to be displayed with --json
-    return { orgId: this.org.getOrgId(), outputString: 'success' };
+    return { orgId: flags['target-org'].getOrgId(), outputString: 'success' };
   }
 }
