@@ -1,19 +1,27 @@
-import { expect, test } from '@salesforce/command/lib/test';
-import { ensureJsonMap, ensureString } from '@salesforce/ts-types';
+import { expect } from 'chai';
+import { TestContext } from '@salesforce/core/lib/testSetup';
+import { stubSfCommandUx } from '@salesforce/sf-plugins-core';
+import VnjikeMetadataLabelUpsert from '../../../../../src/commands/vnjike/metadata/label/upsert';
 
-describe('vnjike:metadata:label:upsert', () => {
-  test
-    .withOrg({ username: 'test@org.com' }, true)
-    .withConnectionRequest(request => {
-      const requestMap = ensureJsonMap(request);
-      if (ensureString(requestMap.url).match(/Organization/)) {
-        return Promise.resolve({ records: [ { Name: 'Super Awesome Org', TrialExpirationDate: '2018-03-20T23:24:11.000+0000'}] });
-      }
-      return Promise.resolve({ records: [] });
-    })
-    .stdout()
-    .command(['vnjike:metadata:label:upsert', '--targetusername', 'scratch2', '-n', 'apiname', '-v', 'value'])
-    .it('runs vnjike:metadata:label:upsert --targetusername test@org.com -n apiname -v value', ctx => {
-      expect(ctx.stdout).to.contain('started');
-    });
+describe('vnjike metadata label upsert', () => {
+  const $$ = new TestContext();
+  let sfCommandStubs: ReturnType<typeof stubSfCommandUx>;
+
+  beforeEach(() => {
+    sfCommandStubs = stubSfCommandUx($$);
+  });
+
+  afterEach(() => {
+    $$.restore();
+  });
+
+  it('should run the upsert command', async () => {
+    const result = await VnjikeMetadataLabelUpsert.run([
+      '--target-org', 'test@org.com', 
+      '--target-label-name', 'apiname', 
+      '--target-label-value', 'value'
+    ]);
+    expect(result).to.be.ok;
+    expect(result.outputString).to.equal('success');
+  });
 });
